:recycle: (#141) 퀴즈 제출 명단 quizAnswer 도메인으로 리팩토링
